Fixes & recordings update
