+++ src/integrationTest/kotlin/com/sourcegraph/cody/AllSuites.kt
@@ -6,4 +6,6 @@ import com.sourcegraph.cody.util.RepeatableSuite
 import org.junit.runner.RunWith
 import org.junit.runners.Suite
 
-@RunWith(RepeatableSuite::class) @Suite.SuiteClasses(ChatTest::class, DocumentCodeTest::class) class AllSuites
+@RunWith(RepeatableSuite::class)
+@Suite.SuiteClasses(ChatTest::class, DocumentCodeTest::class)
+class AllSuites
